requirements.txt
Django>=5.2.4
numpy>=2.3.1
pillow>=11.3.0
scipy>=1.16.0
pandas>=2.3.1
sklearn-som>=1.1.0